# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Course Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /enrollments/{courseId}:
    get:
      operationId: getEnrollment
      tags: [enrollments]
      summary: Get enrollment details for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Enrollment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
...
  /helpdesk/posts:
    get:
      operationId: listPosts
      tags: [helpdesk]
      summary: List help desk posts
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, mine, admin]
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"

    post:
      operationId: createPost
      tags: [helpdesk]
      summary: Create a help desk post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /helpdesk/posts/{id}:
    delete:
      operationId: deletePost
      tags: [helpdesk]
      summary: Delete a helpdesk post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
    patch:
      operationId: updatePostStatus
      tags: [helpdesk]
      summary: Update post status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostStatusUpdate"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    get:
      operationId: getPost
      tags: [helpdesk]
      summary: Get a post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /helpdesk/posts/{id}/comments:
    get:
      operationId: listComments
      tags: [helpdesk]
      summary: List comments for a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"

    post:
      operationId: createComment
      tags: [helpdesk]
      summary: Add a comment to a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
...
  /devices:
    get:
      operationId: listDevices
      tags: [users]
      summary: List active login devices for the current user
      responses:
        "200":
          description: List of devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDevice"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /devices/{id}:
    delete:
      operationId: removeDevice
      tags: [users]
      summary: Remove a login device for the current user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoveDeviceResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /payments:
    post:
      operationId: submitPayment
      tags: [payments]
      summary: Submit a payment for course enrollment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitPaymentInput'
      responses:
        '201':
          description: Payment submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentRecord'
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    get:
      operationId: listPayments
      tags: [payments]
      summary: List all payment submissions (admin)
      responses:
        '200':
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentRecord'
  /payments/{id}/status:
    patch:
      operationId: updatePaymentStatus
      tags: [payments]
      summary: Update payment status (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentStatusUpdate'
      responses:
        '200':
          description: Updated payment record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentRecord'
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  schemas:
...
    Course:
      type: object
      required: [id, title, slug, thumbnail, instructor, category, duration, enrolledCount, isFeatured]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        thumbnail:
          type: string
        instructor:
          $ref: "#/components/schemas/Instructor"
        category:
          type: string
        price:
          type: ["number", "null"]
        duration:
          type: string
        enrolledCount:
          type: integer
        isFeatured:
          type: boolean
        tag:
          type: ["string", "null"]
...
    Enrollment:
      type: object
      required: [id, courseId, course, progress, lastLessonTitle, enrolledAt]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        course:
          $ref: "#/components/schemas/Course"
        progress:
          type: number
        lastLessonTitle:
          type: ["string", "null"]
        enrolledAt:
          type: string
...
    Post:
      type: object
      required: [id, title, content, author, category, status, createdAt, commentCount]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        author:
          $ref: "#/components/schemas/User"
        category:
          type: string
        status:
          type: string
          enum: [open, investigating, acknowledged, resolved]
        createdAt:
          type: string
        commentCount:
          type: integer
        platform:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
    PostStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [open, investigating, acknowledged, resolved]
    PostInput:
      type: object
      required: [title, content, category]
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
    Comment:
      type: object
      required: [id, postId, content, author, createdAt]
      properties:
        id:
          type: integer
        postId:
          type: integer
        content:
          type: string
        author:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
...
    UserDevice:
      type: object
      required: [id, userId, platform, loginAt, lastActiveAt, current]
      properties:
        id:
          type: integer
        userId:
          type: integer
        platform:
          type: string
        userAgent:
          type: ["string", "null"]
        loginAt:
          type: string
          format: date-time
        lastActiveAt:
          type: string
          format: date-time
        current:
          type: boolean
          description: Whether this is the device making the current request
    RemoveDeviceResult:
      type: object
      required: [wasCurrent]
      properties:
        wasCurrent:
          type: boolean
          description: True if the removed device was the caller's own session
    User:
      type: object
      required: [id, name, email]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        studentId:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [student, instructor, admin, mentor, editor]
        addressLine1:
          type: ["string", "null"]
        addressLine2:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        zipCode:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        educationLevel:
          type: ["string", "null"]
        examTitle:
          type: ["string", "null"]
        institute:
          type: ["string", "null"]
        passingYear:
          type: ["string", "null"]
        cvLink:
          type: ["string", "null"]
        githubLink:
          type: ["string", "null"]
        portfolioLink:
          type: ["string", "null"]
        linkedinLink:
          type: ["string", "null"]
        professionalImageLink:
          type: ["string", "null"]
...
    Instructor:
      type: object
      required: [id, name, avatar, title]
      properties:
        id:
          type: integer
        name:
          type: string
        avatar:
          type: string
        title:
          type: string
        bio:
          type: ["string", "null"]
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    PaymentStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, verified, rejected]
    SubmitPaymentInput:
      type: object
      required: [courseName, amount, paymentMethod, paymentPhone, txnId]
      properties:
        courseName:
          type: string
          minLength: 1
        amount:
          type: integer
          minimum: 1
        paymentMethod:
          type: string
          enum: [bkash, nagad]
        paymentPhone:
          type: string
          minLength: 6
        txnId:
          type: string
          minLength: 1
    PaymentRecord:
      type: object
      required: [id, userId, courseName, amount, paymentMethod, paymentPhone, txnId, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        courseName:
          type: string
        amount:
          type: integer
        paymentMethod:
          type: string
        paymentPhone:
          type: string
        txnId:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        userName:
          type: string
          nullable: true
        userEmail:
          type: string
          nullable: true
...